requirements.txt
accelerate==0.27.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.41.2.post2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.0
colorama==0.4.6
datasets==3.5.0
dependency-injector==4.46.0
dill==0.3.8
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
packaging==24.2
pandas==2.2.3
peft==0.7.1
pillow==11.0.0
propcache==0.3.1
protobuf==4.25.3
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.39.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ulid-py==1.1.0
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.0
